requirements.txt
allure-pytest==2.8.22
allure-python-commons==2.8.22
apipkg==1.5
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
execnet==1.7.1
greenlet==0.4.17
idna==2.10
iniconfig==1.1.1
packaging==20.7
playwright==0.162.1
pluggy==0.13.1
py==1.9.0
pyee==8.1.0
pyparsing==2.4.7
pytest==6.1.2
pytest-base-url==1.4.2
pytest-playwright==0.0.9
pytest-xdist==2.1.0
python-slugify==4.0.1
requests==2.25.0
six==1.15.0
slugify==0.0.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
